Fixing some bugs
Bug fixes:
* Not being able to select a part after deleting one
	(Only happened when the mouse hovered over a handle)
* Choosing plane stays enabled when selection changes to {}
	(Plane button disappears, now the plane selection also stops)
* Choosing plane doesn't actually select a plane sometimes
	(Result of MouseLeave not firing always for H_PLANE in Adornments)

A bug that isn't fixed, but that's known:
I fixed the "choosing a plane doesn't actually choose" (point 2 above).
Since I just added a new way of checking if it should be clicked, that works now.
However, the plane-button in the top left corner still uses the MouseLeave event only.
Now, if the bug is meant to happen, the plane gets selected properly, but the button stays blue.
(Also the grid lines stay highlighted grey. Hovering over the button and back will fix it all)
Still, a small visual bug is preferable over a bug in functionality I would say.