Removed waits in rig creation & localized variables
There were a bunch of wait(0.1)s tossed into the rig creation scripts (8
per rig) that didn't seem to serve any purpose and just made the rig
creation take longer than it needed to, so I removed them. Now, clicking
a rig create button creates the rig instantly instead of creating it
piece by piece with a tenth second delay in between each piece.

A lot of the variables were not localized and were giving off the
warning underline so I localized them to remove the warning. I also
changed "Workspace" (now deprecated) to "workspace" to, and while I was
at it I replaced game.Workspace with workspace for consistency's sake
(also because the global workspace variable was created specifically for
that purpose)