Heavy Adornment Memory Optimizations
Theres some memory consumption problem with the Explorer, that happens when multiple objects are being created and destroyed during each frame. This has been a problem for several months now, but it has become more notable with this plugin, due to the CoreGui suddenly becoming visible in the Explorer.

I updated parts of the Adornment module to utilize a smart object pooling system, so that objects don't have to get cleared and regenerated during each frame. As a side effect, this actually makes the tool faster than it was before, and it can handle large parts without the framerate panicking too much.